Python Library For Model Evaluation (#2)
Parent Design: [designs/formak_v0.md](../designs/formak_v0.md)
See Also: [designs/python_ui_demo.md](../designs/python_ui_demo.md)
Status: 2. Feature Tests

## Overview

FormaK aims to combine symbolic modeling for fast, efficient system modelling with code generation to create performant code that is easy to use.

The values (in order) are:

- Easy to use
- Performant

The Five Key Elements the library provides to achieve this (see parent) are:
1. Python Interface to define models
2. Python implementation of the model and supporting tooling
3. Integration to scikit-learn to leverage the model selection and parameter tuning functions
4. C++ and Python to C++ interoperability for performance
5. C++ interfaces to support a variety of model uses

This design provides the initial implementation of second of the Five Keys "Python implementation of the model and supporting tooling". This design also prepares for the third of the 5 Keys: "Integration to scikit-learn to leverage the model selection and parameter tuning functions". At this stage it is helpful to inform the design of the tooling so that it won't have any big hurdles to the next steps in the design.

## Solution Approach

- Tooling
- Transparent Compilation