Scikit-Learn Integration (#3)
Parent Design: [designs/formak_v0.md](../designs/formak_v0.md)
See Also:
[designs/python_library_for_model_evaluation.md](../designs/python_library_for_model_evaluation.md)

##  Overview

FormaK aims to combine symbolic modeling for fast, efficient system
modelling
with code generation to create performant code that is easy to use.

The values (in order) are:

- Easy to use
- Performant

The Five Key Elements the library provides to achieve this (see parent)
are:
1. Python Interface to define models
2. Python implementation of the model and supporting tooling
3. Integration to scikit-learn to leverage the model selection and
parameter tuning functions
4. C++ and Python to C++ interoperability for performance
5. C++ interfaces to support a variety of model uses

This design provides the initial implementation of third of the Five
Keys
"Integration to scikit-learn to leverage the model selection and
parameter
tuning functions". Scikit-learn is a common library who's interface is
replicated many places (e.g. dask-ml for scaling up machine learning
tasks)
that's a good place to start with for an easy to use library.

Why is scikit-learn and machine learning relevant? Conceptually, a
detailed,
physical model derived from first principles describes both one complex
model,
as well as a space of models derived via simplifications, enhancements
or even
disconnected approximations from the original model. Using data from the
system
we hope to describe, we can select the appropriate model from the space.
This
process is very analogous to a machine learning model, where we have one
idea of
how to approximate the system and want to select machine learning models
(in a
more algorithmic sense of the term models) and their parameters to best
fit
data.

### Regression 

From scikit-learn:

"[Regression](https://scikit-learn.org/stable/supervised_learning.html#supervised-learning):
Predicting a continuous-valued attribute associated with an object."

Regression and its use as a predictor of a continuous valued attribute
are the
essential function for the models we hope to evaluate by using
historical data
to predict new values from some prior inputs (model state, control
inputs).

## Solution Approach

The basic step will be to integrate the ui.Model into an interface like
a
scikit-learn model. This will allow for easy integration with
scikit-learn
pipelines, model selection and other functionality.

The key classes involved are:
- `py.Model`: (new) Class encapsulating the model for running a model
efficiently in Python code
- A scikit-learn interface, starting with`.fit`and `.predict`for
regression

The ui library will likely also need some updates to support the
configuration
of the scikit-learn like behavior.

Keeping the machine learning analogy in mind, the key elements from
scikit-learn for this design is regression. This will keep the design
focused
and unlock future advances.