Calibration! (#12)
## Overview

FormaK aims to combine symbolic modeling for fast, efficient system
modelling with code generation to create performant code that is easy to
use.

This design provides an extension to the fifth of the Five Keys "C++
interfaces to support a variety of model uses".

When defining a model, in theory everything can be estimated as a state
or provided as a control input; however, having the ability to provide
calibrations can be helpful or even essential. For example, in the NASA
rocket dataset the position of the IMU on the rocket is calibrated ahead
of time so it doesn't need to be estimated online.

For the case of a suite of ranging sensors (say ultra-wideband time of
flight sensors), the calibration term allows for easily setting up a
single model with different calibrations for the known positions of each
sensor in the reference frame. Without the calibration, each pose would
be arbitrary and require solving a problem beyond what is suited to a
Kalman Filter. With the calibration, the sensor model can be defined
once and then calibrated multiple times at runtime based on how the
sensors are set up for a particular use case.

The Calibration use case also provides additional functionality on top
of the Control inputs. The two categories conceptually overlap as
"known" values that are accepted in the state update; however, the
Calibration values are also available within the sensor model. With just
a State and Control input, the state needs to accept control inputs as a
pass through to sensor models. This adds a compute penalty for
computations with the state.

Supporting calibration is a big step forward in functionality for FormaK
that enables a variety of new model use cases.