FormaK Runtime: "ManagedFilter" (#17)
FormaK aims to combine symbolic modeling for fast, efficient system
modelling with code generation to create performant code that is easy to
use.

The Five Key Elements the library provides to achieve this user
experience are:
1. Python Interface to define models
2. Python implementation of the model and supporting tooling
3. Integration to scikit-learn to leverage the model selection and
parameter tuning functions
4. C++ and Python to C++ interoperability for performance
5. C++ interfaces to support a variety of model uses

This design provides an extension to the second of the Five Keys "Python
implementation of the model..." and the fifth of the Five Keys "C++
interfaces to support a variety of model uses" to support easier to use
filters. The current implementation provides some of the math for the
EKF (process update, sensor update) which can be used in a flexible
manner, but don't necessarily meet the easy to use goal for the project.

This feature aims to make the library easier to use by providing a
managed filter with a single interface.

This feature also has a forward looking benefit: by setting up the
`ManagedFilter` structure now, it will be easier to add "netcode"
features, logging and other runtime benefits within the context of a
unified Managed Filter.

## Solution Approach

The primary class of interest will be a new class `ManagedFilter` (in
both C++ and Python).

This class will have a member function `tick` that will be the primary
user facing to the rest of the filter logic.

```python
    # No sensor reading, process update only
    tick(time)
    tick(time, [])

    # One sensor reading
    tick(time, [reading])

    # Multiple sensor readings
    tick(time, [reading1, reading2, reading3])
```

The function will return the current state and variance of the filter
after processing the tick. By using the `ManagedFilter` wrapper, the
user doesn't need to worry about tracking states, variances, times, the
process model (or
models in the future), sensor models. Instead the user just passes in
the desired output time and any new information from sensors and gets
the result.