Innovation Filtering (#18)
## Overview


FormaK aims to combine symbolic modeling for fast, efficient system
modelling with code generation to create performant code that is easy to
use.

The Five Key Elements the library provides to achieve this user
experience are:
1. Python Interface to define models
2. Python implementation of the model and supporting tooling
3. Integration to scikit-learn to leverage the model selection and
parameter tuning functions
4. C++ and Python to C++ interoperability for performance
5. C++ interfaces to support a variety of model uses

This design focuses on "Integration to scikit-learn to leverage the
model selection and parameter tuning functions". The end result will be
updates to Python and C++ models, but the primary effort will be
improving the scikit-learn tooling. The current implementation provides
some of the math for the EKF (process update, sensor update); however,
that leaves the models open
to spurious readings. In the current implementation, if a FormaK filter
gets a reading that is almost infinite, say 1e300, the response of the
filter will be proportional and the response will cause the EKF to
diverge as one or more of the states jump to positive infinity. This
creates a gap in the filter that violates one of the FormaK values: easy
to use (perhaps in this case it's
helpful to think of this as easy to use *correctly*).

This problem is known as jump detection. As the filter considers a
sensor update the filter measures the divergence of the reading from
what the model expects and then can reject readings with a divergence
that is too large to be acceptable.

How should the filter measure this divergence? The filter can use the
measurement innovation $$\tilde{y} = y - h(x)$$ where y is the
measurement, $h$ is the measurement model and $x$ is the current state
estimate and $\tilde{y}$ is the resulting innovation. Given the
innovation, the definition of too large can be calculated from the
expected covariance:
$$\tilde{y}^{T} C^{-1} \tilde{y} - m > k \sqrt{2m}$$ 
where $C$ is the expected covariance of the measurement, $m$ is the
dimension of the measurement vector and $k$ is the editing threshold. If
this inequality is true, then the measurement is "too large" and should
be filtered (edited). [1]

[1] This approach to innovation filtering, referred to as editing in the
text, is adapted from "Advanced Kalman Filtering, Least Squares and
Modeling" by Bruce P. Gibbs (referred to as [1] or AKFLSM)

Innovations take advantage of the property that errors are white
(normally distributed) when all models are correct and, when operating
in steady state, most variation in the sensor reading is expected to
come from the noise in the sensor instead of noise in the state.
Intuitively, if the sensor is consistently providing low variance
updates, the state should converge further based on that information
until the noise in the state relative to the sensor is low.

By implementing innovation filtering, the filter will become much more
robust to errors in the sensor, the model and the overall system. By
hooking into values that are already calculated as part of the sensor
update there should be minimal additional performance cost.