Hyperparameter Selection (#21)
Author: Buck Baskin [@buck@fosstodon.org](https://fosstodon.org/@buck)
Created: 2023-11-08
Updated: 2023-11-08
Parent Design:
[designs/sklearn-integration.md](../designs/sklearn-integration.md)
See Also:
[designs/python_library_for_model_evaluation.md](../designs/python_library_for_model_evaluation.md),
[designs/innovation_filtering.md](../designs/innovation_filtering.md)


## Overview

FormaK aims to combine symbolic modeling for fast, efficient system
modelling with code generation to create performant code that is easy to
use.

The Five Key Elements the library provides to achieve this user
experience are:
1. Python Interface to define models
2. Python implementation of the model and supporting tooling
3. Integration to scikit-learn to leverage the model selection and
parameter tuning functions
4. C++ and Python to C++ interoperability for performance
5. C++ interfaces to support a variety of model uses

This design focuses on "Integration to scikit-learn to leverage the
model selection and parameter tuning functions". More specifically, this
design focuses on using scikit-learn tooling to automatically select the
innovation filtering level from data.

The promise of this design is that all parameters could be selected
automatically based on data instead of requiring hand tuning; however,
this design will focus narrowly on selecting the innovation filtering
level as a motivating example.